Write permissions now checked and handled correctly
When users attempt to upload fasta/genbank/embl files, vectormaps or
generate pdfs a check is performed if the directories are writeable and
an appropriate error is thrown in case they are not !

+++ application/models/plasmid_model.php
@@ -191,8 +191,15 @@ class Plasmid_model extends CI_Model {
 
 	public function upload_vectormap($plasmid_id)
 	{
+			//check if directory is writeable
+			if(!is_writable(realpath(APPPATH . '../img/vectormaps')))
+			{
+				$_SESSION['error'] = "Directory img/vectormaps not writeable, contact the webmaster to set this up.";
+				return false;
+			}
 			//session should be started by controller
 			ob_start();
+		
 		
 		    $config =  array (
 				'allowed_types' => 'gif|png|jpg|jpeg',
@@ -243,6 +250,11 @@ class Plasmid_model extends CI_Model {
 	
 	public function upload_sequence()
 	{
+			if(!is_writable(realpath(APPPATH . '../tmp')))
+			{
+				$_SESSION['error'] = "Directory tmp not writeable, contact the webmaster to set this up.";
+				return false;
+			}
 			//session should be started by controller
 			ob_start();		
 			
@@ -285,6 +297,13 @@ class Plasmid_model extends CI_Model {
 	public function upload_genbank()
 	{
 			$this->load->helper('genbank');
+			
+			if(!is_writable(realpath(APPPATH . '../tmp')))
+			{
+				$_SESSION['error'] = "Directory tmp not writeable, contact the webmaster to set this up.";
+				return false;
+			}
+			
 			//session should be started by controller
 			ob_start();
 		
@@ -326,6 +345,13 @@ class Plasmid_model extends CI_Model {
 	public function upload_embl()
 	{
 			$this->load->helper('embl');
+			
+			if(!is_writable(realpath(APPPATH . '../tmp')))
+			{
+				$_SESSION['error'] = "Directory tmp not writeable, contact the webmaster to set this up.";
+				return false;
+			}
+			
 			//session should be started by controller
 			ob_start();
 		
@@ -388,16 +414,22 @@ class Plasmid_model extends CI_Model {
 
 	public function remove_vectormap($plasmid_id)
 	{
-				$img_path = realpath(APPPATH . '../img/vectormaps/');
-				$img_path = $img_path . "/" . $plasmid_id; 
-				if (file_exists($img_path. ".png")) {unlink($img_path. ".png"); }
-				if (file_exists($img_path. "_thumb.png")) {unlink($img_path. "_thumb.png"); }
-				if (file_exists($img_path. ".jpg")) {unlink($img_path. ".jpg"); }
-				if (file_exists($img_path. "_thumb.jpg")) {unlink($img_path. "_thumb.jpg"); }
-				if (file_exists($img_path. ".jpeg")) {unlink($img_path. ".jpeg"); }
-				if (file_exists($img_path. "_thumb.jpeg")) {unlink($img_path. "_thumb.jpeg"); }
-				if (file_exists($img_path. ".gif")) {unlink($img_path. ".gif"); }
-				if (file_exists($img_path. "_thumb.gif")) {unlink($img_path. "_thumb.gif"); }		
+		if(!is_writable(realpath(APPPATH . '../img/vectormaps/')))
+		{
+			$_SESSION['error'] = "Directory img/vectormaps/ not writeable, contact the webmaster to set this up.";
+			return false;
+		}
+		
+		$img_path = realpath(APPPATH . '../img/vectormaps/');
+		$img_path = $img_path . "/" . $plasmid_id; 
+		if (file_exists($img_path. ".png")) {unlink($img_path. ".png"); }
+		if (file_exists($img_path. "_thumb.png")) {unlink($img_path. "_thumb.png"); }
+		if (file_exists($img_path. ".jpg")) {unlink($img_path. ".jpg"); }
+		if (file_exists($img_path. "_thumb.jpg")) {unlink($img_path. "_thumb.jpg"); }
+		if (file_exists($img_path. ".jpeg")) {unlink($img_path. ".jpeg"); }
+		if (file_exists($img_path. "_thumb.jpeg")) {unlink($img_path. "_thumb.jpeg"); }
+		if (file_exists($img_path. ".gif")) {unlink($img_path. ".gif"); }
+		if (file_exists($img_path. "_thumb.gif")) {unlink($img_path. "_thumb.gif"); }		
 	}
 
 	public function remove_embl($plasmid_id)
